pyproject.toml
[project]
name = "twilio-whatsapp"
version = "0.1.0"
description = "WhatsApp chatbot using Twilio and Google Gemini LLM for data engineering expertise"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "twilio>=9.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "loguru>=0.7.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
]
